+++ src/components/component-interfaces.ts
@@ -1,3 +1,9 @@
+import type { ChangeEvent } from 'react';
+
 export interface ComponentAnimationProps {
   useAnimation?: boolean;
+}
+
+export interface BaseEvent<T extends HTMLElement> {
+  originalEvent?: ChangeEvent<T>;
 }